# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateRestaurantRequest:
    properties:
      name:
        type: string
      plan:
        type: string
    required:
    - name
    type: object
  dto.CreateTableRequest:
    properties:
      number:
        type: integer
    required:
    - number
    type: object
  dto.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  dto.RestaurantResponse:
    properties:
      id:
        type: string
      name:
        type: string
      plan:
        type: string
    type: object
...
  dto.TableResponse:
    properties:
      id:
        type: string
      number:
        type: integer
      qr_code:
        type: string
      restaurant_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Sistema de llamado de mozos para restaurantes
  title: Waiter API
  version: "1.0"
paths:
...
  /api/v1/admin/tables/{id}/regenerate-qr:
    post:
      description: Genera un nuevo código QR aleatorio para una mesa (admin)
      parameters:
      - description: ID de la mesa
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - AdminToken: []
      summary: Regenerar QR de mesa
      tags:
      - tables
...
  /api/v1/restaurants:
    get:
      description: Retorna todos los restaurantes del sistema (solo superadmin)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.RestaurantResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - AdminToken: []
      summary: Listar todos los restaurantes
      tags:
      - restaurants
    post:
      consumes:
      - application/json
      description: Crea un nuevo restaurante
      parameters:
      - description: Datos del restaurante
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRestaurantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RestaurantResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - AdminToken: []
      summary: Crear restaurante
      tags:
      - restaurants
  /api/v1/restaurants/{restaurantId}:
    get:
      description: Retorna un restaurante por su ID
      parameters:
      - description: ID del restaurante
        in: path
        name: restaurantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RestaurantResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - AdminToken: []
      summary: Obtener restaurante
      tags:
      - restaurants
...
  /api/v1/restaurants/{restaurantId}/tables:
    get:
      description: Retorna las mesas de un restaurante
      parameters:
      - description: ID del restaurante
        in: path
        name: restaurantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TableResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - AdminToken: []
      summary: Listar mesas
      tags:
      - tables
    post:
      consumes:
      - application/json
      description: Crea una nueva mesa para un restaurante
      parameters:
      - description: ID del restaurante
        in: path
        name: restaurantId
        required: true
        type: string
      - description: Datos de la mesa
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTableRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TableResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - AdminToken: []
      summary: Crear mesa
      tags:
      - tables
...
securityDefinitions:
  AdminToken:
    description: 'JWT de administrador. Formato: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
  SessionToken:
    description: 'JWT de sesión de cliente (QR). Formato: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"